simplify the mental model for mixing bytes/u32/u64
By changing the refill case for read_u32 and read_u64 from "refill a bit
early and then proceed as usual" into treating it like a normal 4-byte
read, the reproducibility promise gets a lot simpler. It also doesn't
seem to have any significant performance cost in microbenchmarks. If
anything, it might be *better* in real-world usage because it finally
gave me a way to tell the compiler that the refill case is cold.