stop exposing backend choice to the benchmark
The main motivation is to get rid of the unstable_internals feature,
which ideally shouldn't exist at all. It also makes the powerset of
features to test (with a naive `cargo hack` invocation) even larger,
which has started to become annoying with the new rand 0.9 feature.

The ability to easily benchmark non-default features is lost for now,
but this should be fine as long as I'm not making major changes to them.
In the future, I might add some internal-use-only `cfg`s to force a
particular backend. But no need to figure out the details yet.